Fix order items
